refactor: fix rate limit middleware file name
